+++ src/Field.php
@@ -387,6 +387,12 @@ class Field extends HtmlField implements ElementContainerFieldInterface, Mergeab
     public ?int $wordLimit = null;
 
     /**
+     * @var int|null The total number of characters allowed.
+     * @since 4.8.0
+     */
+    public ?int $characterLimit = null;
+
+    /**
      * @var bool Whether the word count should be shown below the field.
      * @since 3.2.0
      */
@@ -460,6 +466,15 @@ class Field extends HtmlField implements ElementContainerFieldInterface, Mergeab
             $config['sourceEditingGroups'] = null;
         }
 
+        if (isset($config['limitUnit'], $config['fieldLimit'])) {
+            if ($config['limitUnit'] === 'chars') {
+                $config['characterLimit'] = (int)$config['fieldLimit'] ?: null;
+            } else {
+                $config['wordLimit'] = (int)$config['fieldLimit'] ?: null;
+            }
+            unset($config['limitUnit'], $config['fieldLimit']);
+        }
+
         if (isset($config['entryTypes']) && $config['entryTypes'] === '') {
             $config['entryTypes'] = [];
         }
@@ -477,6 +492,9 @@ class Field extends HtmlField implements ElementContainerFieldInterface, Mergeab
         if ($this->wordLimit === 0) {
             $this->wordLimit = null;
         }
+        if ($this->characterLimit === 0) {
+            $this->characterLimit = null;
+        }
     }
 
     /**
@@ -486,6 +504,7 @@ class Field extends HtmlField implements ElementContainerFieldInterface, Mergeab
     {
         return array_merge(parent::defineRules(), [
             ['wordLimit', 'number', 'min' => 1],
+            ['characterLimit', 'number', 'min' => 1],
         ]);
     }
 
@@ -496,7 +515,22 @@ class Field extends HtmlField implements ElementContainerFieldInterface, Mergeab
     {
         $rules = [];
 
-        if ($this->wordLimit) {
+        if ($this->characterLimit) {
+            $rules[] = [
+                function(ElementInterface $element) {
+                    $value = strip_tags((string)$element->getFieldValue($this->handle));
+                    if (strlen($value) > $this->characterLimit) {
+                        $element->addError(
+                            "field:$this->handle",
+                            Craft::t('ckeditor', '{field} should contain at most {max, number} {max, plural, one{character} other{characters}}.', [
+                                'field' => Craft::t('site', $this->name),
+                                'max' => $this->characterLimit,
+                            ]),
+                        );
+                    }
+                },
+            ];
+        } elseif ($this->wordLimit) {
             $rules[] = [
                 function(ElementInterface $element) {
                     $value = html_entity_decode((string)$element->getFieldValue($this->handle));
@@ -1038,7 +1072,7 @@ JS;
         $uiLanguage = BaseCkeditorPackageAsset::uiLanguage();
         $uiTranslationImport = "import coreTranslations from 'ckeditor5/translations/$uiLanguage.js';";
 
-        $view->registerScriptWithVars(fn($baseConfigJs, $toolbarJs, $languageJs, $showWordCountJs, $wordLimitJs) => <<<JS
+        $view->registerScriptWithVars(fn($baseConfigJs, $toolbarJs, $languageJs, $showWordCountJs, $wordLimitJs, $characterLimitJs) => <<<JS
 $imports
 $uiTranslationImport
 import {create} from '@craftcms/ckeditor';
@@ -1085,6 +1119,15 @@ import {create} from '@craftcms/ckeditor';
           container.removeClass('error warning');
         }
       }
+      if ($characterLimitJs) {
+        if (stats.characters > $characterLimitJs) {
+          container.addClass('error');
+        } else if (stats.characters >= Math.floor($characterLimitJs * .9)) {
+          container.addClass('warning');
+        } else {
+          container.removeClass('error warning');
+        }
+      }
       onUpdate(stats);
     };
   } else {
@@ -1106,6 +1149,7 @@ JS,
                 ],
                 $this->showWordCount,
                 $this->wordLimit ?: 0,
+                $this->characterLimit ?: 0,
             ],
             View::POS_END,
             ['type' => 'module']